drivers/hpe: Add config to suppress SSL warnings
This commit adds a new config, hpe3par_suppress_ssl_warnings, to
suppress the WSAPI HTTPS SSL certificate verification warnings.

This config defaults to `True` because the HPE driver currently lacks
the support for specifying the server SSL certificate to be used for
verification and, for all practical purposes, the verification step is
unnecessary because both 3PAR controller and cinder service will be
located in a secure isolated network.

Upstream-Status: Pending
Signed-off-by: Stephanos Ioannidis <root@stephanos.io>